Some formatting fixes
