+++ src/main/java/spark/Request.java
@@ -33,10 +33,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import spark.routematch.RouteMatch;
-import spark.utils.urldecoding.UrlDecode;
 import spark.utils.IOUtils;
 import spark.utils.SparkUtils;
 import spark.utils.StringUtils;
+import spark.utils.urldecoding.UrlDecode;
 
 /**
  * Provides information about the HTTP request
@@ -44,28 +44,28 @@ import spark.utils.StringUtils;
  * @author Per Wendel
  */
 public class Request {
-    
+
     private static final org.slf4j.Logger LOG = org.slf4j.LoggerFactory.getLogger(Request.class);
-    
+
     private static final String USER_AGENT = "user-agent";
-    
+
     private Map<String, String> params;
     private List<String> splat;
     private QueryParamsMap queryMap;
-    
+
     private HttpServletRequest servletRequest;
-    
+
     private Session session = null;
     private boolean validSession = false;
     private String matchedPath = null;
-    
-    
+
+
     /* Lazy loaded stuff */
     private String body = null;
     private byte[] bodyAsBytes = null;
-    
+
     private Set<String> headers = null;
-    
+
     //    request.body              # request body sent by the client (see below), DONE
     //    request.scheme            # "http"                                DONE
     //    request.path_info         # "/foo",                               DONE
@@ -88,11 +88,11 @@ public class Request {
     //    request.script_name       # "/example"
     //    request.form_data?        # false
     //    request.referrer          # the referrer of the client or '/'
-    
+
     protected Request() {
         // Used by wrapper
     }
-    
+
     /**
      * Constructor
      *
@@ -104,7 +104,7 @@ public class Request {
         this.matchedPath = match.getMatchUri();
         changeMatch(match);
     }
-    
+
     /**
      * Constructor - Used to create a request and no RouteMatch is available.
      *
@@ -112,21 +112,21 @@ public class Request {
      */
     Request(HttpServletRequest request) {
         this.servletRequest = request;
-        
+
         // Empty
         params = new HashMap<>();
         splat = new ArrayList<>();
     }
-    
+
     protected void changeMatch(RouteMatch match) {
         List<String> requestList = SparkUtils.convertRouteToList(match.getRequestURI());
         List<String> matchedList = SparkUtils.convertRouteToList(match.getMatchUri());
-        
+
         this.matchedPath = match.getMatchUri();
         params = getParams(requestList, matchedList);
         splat = getSplat(requestList, matchedList);
     }
-    
+
     /**
      * Returns the map containing all route params
      *
@@ -135,7 +135,7 @@ public class Request {
     public Map<String, String> params() {
         return Collections.unmodifiableMap(params);
     }
-    
+
     /**
      * Returns the value of the provided route pattern parameter.
      * Example: parameter 'name' from the following pattern: (get '/hello/:name')
@@ -147,57 +147,57 @@ public class Request {
         if (param == null) {
             return null;
         }
-        
+
         if (param.startsWith(":")) {
             return params.get(param.toLowerCase()); // NOSONAR
         } else {
             return params.get(":" + param.toLowerCase()); // NOSONAR
         }
     }
-    
+
     /**
      * @return an array containing the splat (wildcard) parameters
      */
     public String[] splat() {
         return splat.toArray(new String[splat.size()]);
     }
-    
+
     /**
      * @return request method e.g. GET, POST, PUT, ...
      */
     public String requestMethod() {
         return servletRequest.getMethod();
     }
-    
+
     /**
      * @return the scheme
      */
     public String scheme() {
         return servletRequest.getScheme();
     }
-    
+
     /**
      * @return the host
      */
     public String host() {
         return servletRequest.getHeader("host");
     }
-    
+
     /**
      * @return the user-agent
      */
     public String userAgent() {
         return servletRequest.getHeader(USER_AGENT);
     }
-    
+
     /**
      * @return the server port
      */
     public int port() {
         return servletRequest.getServerPort();
     }
-    
-    
+
+
     /**
      * @return the path info
      * Example return: "/example/foo"
@@ -205,67 +205,69 @@ public class Request {
     public String pathInfo() {
         return servletRequest.getPathInfo();
     }
-    
+
     /**
      * @return the matched route
      * Example return: "/account/:accountId"
      */
-    public String matchedPath() { return this.matchedPath; }
-    
+    public String matchedPath() {
+        return this.matchedPath;
+    }
+
     /**
      * @return the servlet path
      */
     public String servletPath() {
         return servletRequest.getServletPath();
     }
-    
+
     /**
      * @return the context path
      */
     public String contextPath() {
         return servletRequest.getContextPath();
     }
-    
+
     /**
      * @return the URL string
      */
     public String url() {
         return servletRequest.getRequestURL().toString();
     }
-    
+
     /**
      * @return the content type of the body
      */
     public String contentType() {
         return servletRequest.getContentType();
     }
-    
+
     /**
      * @return the client's IP address
      */
     public String ip() {
         return servletRequest.getRemoteAddr();
     }
-    
+
     /**
      * @return the request body sent by the client
      */
     public String body() {
-        
+
         if (body == null) {
             body = StringUtils.toString(bodyAsBytes(), servletRequest.getCharacterEncoding());
         }
-        
+
         return body;
     }
-    
+
     public byte[] bodyAsBytes() {
         if (bodyAsBytes == null) {
             readBodyAsBytes();
         }
         return bodyAsBytes;
     }
-    
+
     private void readBodyAsBytes() {
         try {
             bodyAsBytes = IOUtils.toByteArray(servletRequest.getInputStream());
@@ -273,14 +275,14 @@ public class Request {
             LOG.warn("Exception when reading body", e);
         }
     }
-    
+
     /**
      * @return the length of request.body
      */
     public int contentLength() {
         return servletRequest.getContentLength();
     }
-    
+
     /**
      * Gets the query param
      *
@@ -291,9 +293,9 @@ public class Request {
     public String queryParams(String queryParam) {
         return servletRequest.getParameter(queryParam);
     }
-    
+
     //CS304 Issue link:https://github.com/perwendel/spark/issues/1061
-    
+
     /**
      * Gets the query param and encode it
      *
@@ -302,9 +304,9 @@ public class Request {
      * Example: query parameter 'me' from the URI: /hello?id=fool.
      */
     public String queryParamsSafe(final String queryParam) {
-        return EnDeCode.encode(servletRequest.getParameter(queryParam));
+        return Base64.encode(servletRequest.getParameter(queryParam));
     }
-    
+
     /**
      * Gets the query param, or returns default value
      *
@@ -317,7 +319,7 @@ public class Request {
         String value = queryParams(queryParam);
         return value != null ? value : defaultValue;
     }
-    
+
     /**
      * Gets all the values of the query param
      * Example: query parameter 'id' from the following request URI: /hello?id=foo&amp;id=bar
@@ -328,7 +330,7 @@ public class Request {
     public String[] queryParamsValues(String queryParam) {
         return servletRequest.getParameterValues(queryParam);
     }
-    
+
     /**
      * Gets the value for the provided header
      *
@@ -338,14 +340,14 @@ public class Request {
     public String headers(String header) {
         return servletRequest.getHeader(header);
     }
-    
+
     /**
      * @return all query parameters
      */
     public Set<String> queryParams() {
         return servletRequest.getParameterMap().keySet();
     }
-    
+
     /**
      * @return all headers
      */
@@ -359,14 +361,14 @@ public class Request {
         }
         return headers;
     }
-    
+
     /**
      * @return the query string
      */
     public String queryString() {
         return servletRequest.getQueryString();
     }
-    
+
     /**
      * Sets an attribute on the request (can be fetched in filters/routes later in the chain)
      *
@@ -376,7 +378,7 @@ public class Request {
     public void attribute(String attribute, Object value) {
         servletRequest.setAttribute(attribute, value);
     }
-    
+
     /**
      * Gets the value of the provided attribute
      *
@@ -388,8 +390,8 @@ public class Request {
     public <T> T attribute(String attribute) {
         return (T) servletRequest.getAttribute(attribute);
     }
-    
-    
+
+
     /**
      * @return all attributes
      */
@@ -401,23 +403,23 @@ public class Request {
         }
         return attrList;
     }
-    
+
     /**
      * @return the raw HttpServletRequest object handed in by Jetty
      */
     public HttpServletRequest raw() {
         return servletRequest;
     }
-    
+
     /**
      * @return the query map
      */
     public QueryParamsMap queryMap() {
         initQueryMap();
-        
+
         return queryMap;
     }
-    
+
     /**
      * @param key the key
      * @return the query map
@@ -425,13 +427,13 @@ public class Request {
     public QueryParamsMap queryMap(String key) {
         return queryMap().get(key);
     }
-    
+
     private void initQueryMap() {
         if (queryMap == null) {
             queryMap = new QueryParamsMap(raw());
         }
     }
-    
+
     /**
      * Returns the current session associated with this request,
      * or if the request does not have a session, creates one.
@@ -445,7 +447,7 @@ public class Request {
         }
         return session;
     }
-    
+
     /**
      * Returns the current session associated with this request, or if there is
      * no current session and <code>create</code> is true, returns  a new session.
@@ -467,7 +469,7 @@ public class Request {
         }
         return session;
     }
-    
+
     /**
      * @return request cookies (or empty Map if cookies aren't present)
      */
@@ -481,7 +483,7 @@ public class Request {
         }
         return result;
     }
-    
+
     /**
      * Gets cookie by name.
      *
@@ -499,58 +501,58 @@ public class Request {
         }
         return null;
     }
-    
+
     /**
      * @return the part of this request's URL from the protocol name up to the query string in the first line of the HTTP request.
      */
     public String uri() {
         return servletRequest.getRequestURI();
     }
-    
+
     /**
      * @return Returns the name and version of the protocol the request uses
      */
     public String protocol() {
         return servletRequest.getProtocol();
     }
-    
+
     private static Map<String, String> getParams(List<String> request, List<String> matched) {
         Map<String, String> params = new HashMap<>();
-        
+
         for (int i = 0; (i < request.size()) && (i < matched.size()); i++) {
             String matchedPart = matched.get(i);
-            
+
             if (SparkUtils.isParam(matchedPart)) {
-                
+
                 String decodedReq = UrlDecode.path(request.get(i));
-                
+
                 LOG.debug("matchedPart: "
-                          + matchedPart
-                          + " = "
-                          + decodedReq);
-                
+                                  + matchedPart
+                                  + " = "
+                                  + decodedReq);
+
                 params.put(matchedPart.toLowerCase(), decodedReq);
             }
         }
         return Collections.unmodifiableMap(params);
     }
-    
+
     private static List<String> getSplat(List<String> request, List<String> matched) {
         int nbrOfRequestParts = request.size();
         int nbrOfMatchedParts = matched.size();
-        
+
         boolean sameLength = (nbrOfRequestParts == nbrOfMatchedParts);
-        
+
         List<String> splat = new ArrayList<>();
-        
+
         for (int i = 0; (i < nbrOfRequestParts) && (i < nbrOfMatchedParts); i++) {
-            
+
             String matchedPart = matched.get(i);
-            
+
             if (SparkUtils.isSplat(matchedPart)) {
-                
+
                 StringBuilder splatParam = new StringBuilder(request.get(i));
-                
+
                 if (!sameLength && (i == (nbrOfMatchedParts - 1))) {
                     for (int j = i + 1; j < nbrOfRequestParts; j++) {
                         splatParam.append("/");
@@ -564,10 +566,10 @@ public class Request {
                 }
             }
         }
-        
+
         return Collections.unmodifiableList(splat);
     }
-    
+
     /**
      * Set the session validity
      *
@@ -576,5 +578,5 @@ public class Request {
     void validSession(boolean validSession) {
         this.validSession = validSession;
     }
-    
+
 }
