+++ src/test/java/spark/RequestTest.java
@@ -63,7 +63,6 @@ public class RequestTest {
         request = new Request(match, servletRequest);
 
     }
-    
 
     @Test
     public void queryParamShouldReturnsParametersFromQueryString() {
